+++ src/app/(user)/reservation-status/page.tsx
@@ -2,6 +2,16 @@ import ReservationStatus from "@/app/(user)/reservation-status/_components/Reser
 import { getMyActivities } from "@/app/(user)/reservation-status/utils/reservationStatus";
 import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query";
 
+export const metadata = {
+  title: "예약 현황",
+  description: "내 체험의 예약 현황(승인, 거절, 완료)을 표시합니다.",
+  openGraph: {
+    title: "예약 현황 페이지",
+    description: "내 체험의 예약 현황(승인, 거절, 완료)을 표시합니다.",
+    url: "https://travel-kidnap-maker.vercel.app/reservation-status",
+  },
+};
+
 async function ReservationStatusPage() {
   const queryClient = new QueryClient();
